% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMediation.R
\name{BayesMediation}
\alias{BayesMediation}
\title{BayesMediation}
\usage{
BayesMediation(
  Gamma_hat,
  Sd_hat,
  init = "Random",
  iter = 6000,
  warmup = 3000,
  second = F,
  inv = FALSE,
  cor = NULL,
  Raw = T,
  total = F,
  indirect = F
)
}
\arguments{
\item{Gamma_hat}{The estimated exposure and outcome effects gamma_hat}

\item{Sd_hat}{The standarded errors of gamma_hat}

\item{init}{Starting value for gibbs sampler. Either "Random" or "EM"}

\item{iter}{The number of iterations for gibbs sampler}

\item{warmup}{The nnumber of warmup periods}

\item{second}{Whether to run the second stage. Default is False.}

\item{inv}{When inv = False, we are estiming B in Gamma = (I + B) alpha;
inv = True, we are estimating B in Gamma = B Gamma + alpha.}

\item{cor}{The correlation matrix}

\item{Raw}{Whether to include the unprocessed raw outputs. Default
is F}

\item{total}{Whether to include the total effects. Default is F}

\item{indirect}{Whether to include the indirect effects. Default is F.}
}
\value{
A list with elements
\item{summary_first} The summary table of first stage
\item{summary_second} The summary table of second stage (included if second  = T)
\item{total_effect_first} The total effect of each exposure on the outcome,
computed by first stage (included if total = T)
\item{total_effect_second} The total effect of each exposure on the outcome,
computed by second stage (included if total = T and second = T)
\item{indirect_effect_first} The indirect effect of each exposure on the outcome,
computed by first stage (included if indirect = T)
\item{indirect_effect_second} The indirect effect of each exposure on the outcome,
computed by second stage (included if indirect = T and  second = T)
\item{raw_first} The unprocessed output from the first stage (included if raw= T)
\item{raw_second} The unprocessed output from the second stage (included if raw = T and second = T)
}
\description{
Bayes approach for MR Mediation model.
}
