% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\alias{data}
\alias{pos}
\title{This is the data for examples}
\usage{
data(example_data)
}
\description{
\itemize{
  \item data. phenotype file. 1st column is ID, 2nd column is continuous outcome, 3rd column is binary outcome, 4th column is exposure, 5th column is age, 6th column is gender, 7th-last columns are CpGs
  \item pos. CpG locations from the defined region and they are from the same chromosome.
}
}
