% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler_with_corr}
\alias{gibbs_sampler_with_corr}
\title{Perform Gibbs sampling for mediation MR with correlation}
\usage{
gibbs_sampler_with_corr(
  Y,
  Sd_hat,
  trait_corr,
  N,
  B,
  sigma,
  sigma1,
  sigma0,
  p,
  A,
  Z,
  alpha_B,
  beta_B,
  alpha_0,
  alpha_1,
  beta_0,
  beta_1,
  a,
  b,
  Lambda3,
  Lambda_inv3
)
}
\arguments{
\item{Y}{KxP arma::matrix of GWAS marginal associations}

\item{Sd_hat}{KxP arma::matrix of GWAS std errors}

\item{trait_corr}{KxK arma::matrix of correlation of noises.}

\item{N}{number of samples to obtain}

\item{B}{initial value of B arma::matrix}

\item{sigma}{Initial value of prior std deviation for values of B}

\item{sigma1}{K-arma::vector of Initial values of prior slab SDs for each trait's pleiotropy}

\item{sigma0}{K-arma::vector of Initial values of prior splike SDs for each trait's pleiotropy}

\item{p}{K-arma::vector of initial values for p, which is slab probability for mixture}

\item{A}{KxP arma::matrix of initial values for A (ignored)}

\item{Z}{KxP arma::matrix of initial values for Z (ignored), which is latent varaible indicating if A[i,j] comes from slab}

\item{alpha_B}{Shape parameter of InvGamma prior for sigma (for B)}

\item{beta_B}{Scale parameter of InvGamma prior for sigma (for B)}

\item{alpha_0}{K-arma::vector Shape parameter of InvGamma prior for sigma0 (Spike)}

\item{alpha_1}{K-arma::vector Shape parameter of InvGamma prior for sigma0 (Slab)}

\item{beta_0}{K-arma::vector Scale parameter of InvGamma prior for sigma0 (Spike)}

\item{beta_1}{K-arma::vector Scale parameter of InvGamma prior for sigma0 (Slab)}

\item{a}{K-arma::vector First parameter of beta prior for p}

\item{b}{K-arma::vector Second parameter of beta prior for p}

\item{Lambda3}{Precomputed covariance matrix}

\item{Lambda_inv3}{Precomputed inverse covariance matrix}
}
\description{
Model is that Y = (I+B)A + noise
}
