% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{gibbs_wrapper}
\alias{gibbs_wrapper}
\title{Gibbs sampler}
\usage{
gibbs_wrapper(
  N,
  warmup,
  chains = 4,
  Gamma_hat,
  Sd_hat,
  sigma,
  sigma_1,
  sigma_0,
  p,
  parallel = TRUE,
  init = "Random",
  ratio = 10,
  cor_mat = NULL,
  Lambda = NULL,
  Lambda_inv = NULL
)
}
\arguments{
\item{N}{Number of total iterations}

\item{warmup}{Number of warm-up iterations.}

\item{chains}{Number of Chains.}

\item{Gamma_hat}{The observed effects}

\item{Sd_hat}{The standard deviation of Gamma_hat}

\item{sigma}{The prior mean for sigma}

\item{sigma_1}{The prior mean for sigma_1}

\item{sigma_0}{The prior mean for sigma_0}

\item{p}{The prior mean for p}

\item{parallel}{Indicating whether to parallelly run all the chains. 
Default is True.}

\item{init}{Method of Initialization. Must be one of "EM" or "Random". Default
is "Random"}

\item{ratio}{The ratio of initial sigma1 and sigma0 when the initialization method
is "Random". Default is 10.}

\item{cor_mat}{The correlation matrix for noise. Default is null (no correlation)}

\item{Lambda}{The block diagonal matrix such that each block is the covariance 
matrix of the noise for a SNP. Default is NULL (not provided)}

\item{Lambda_inv}{The block diagonal matrix such that each block is the inverse of 
covariance matrix of the noise for a SNP. Default is NULL (not provided).}
}
\value{
A list of which the elements are outputs from each chain. Each of these
outputs is a list with attributes
\item{B}
\item{p}
\item{sigma1}
\item{sigma0}
\item{pi}
\item{sigma}
}
\description{
Given the observed effect and standard error, run gibbs sampler to draw 
samples from the posterior density
}
