\name{MrMediation-package}
\alias{MrMediation-package}
\alias{MrMediation}
\docType{package}
\title{
\packageTitle{MrMediation}
}
\description{
\packageDescription{MrMediation}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MrMediation}
\packageIndices{MrMediation}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MrMediation}

Maintainer: \packageMaintainer{MrMediation}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\examples{
# simple examples of the most important functions ~~
}
