% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{sigma_objective}
\alias{sigma_objective}
\title{The log likelihood function in the M-step}
\usage{
sigma_objective(sigma.sq, weights, values, std.errors)
}
\description{
The log likelihood function in the M-step
}
\keyword{internal}
