% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary_gibbs}
\alias{summary_gibbs}
\title{Produce a summary table of the output from gibbs_wrapper.}
\usage{
summary_gibbs(result, pars, K, T = 1000, inv = FALSE)
}
\arguments{
\item{result}{The output of gibbs_wrapper or gibbs_wrapper_cor}

\item{pars}{A vector of parameters to be included in the summary table}

\item{K}{The number of mediation layers}

\item{T}{The warmup period. Default is 1000.}

\item{inv}{When inv is TRUE, the function shows summary statistics for B in
Gamma = B Gamma + alpha, otherwise it shows summary statistics for B in
Gamma = (I + B) alpha. Default is False}
}
\value{
A table where each row corresponds to a parameter. It has columns:
\item{mean} The average of samples after warm-up
\item{variance} The variance of samples after warm-up
\item{sd} The standard deviation of samples after warm-up
\item{2.5\%} The 2.5\% quantile of samples after warm-up
\item{50\%} The 50% quantile of samples after warm-up
\item{97.5\%} The 97.5\% quantile of samples after warm-up
\item{ESS} The Effective sample sizes (combined from all the chains)
\item{R_hat} A measure of convergence. A value below 1.1 indicates sign
of convergence
}
\description{
Produce a summary table of the output from gibbs_wrapper.
}
