% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_effects.R
\name{total_effect}
\alias{total_effect}
\title{Total effect}
\usage{
total_effect(results, K, warmup = 3000)
}
\arguments{
\item{results}{results from gibbs_wrapper}

\item{K}{number of phenotypes}

\item{warump}{The length of warmup period. Default is 3000}
}
\value{
A matrix of total effect and its quantiles.
}
\description{
Estimate the total effect of an exposure on the outcome
}
