% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot}
\alias{traceplot}
\title{Produce traceplot for the specified parameter}
\usage{
traceplot(results, par, ind, chains = -1, ylim = NA, T = 1000)
}
\arguments{
\item{results}{The output of gibbs_wrapper.}

\item{ind}{The index of the parameter. For "B" it is a two-vector. For "sigma1",
"sigma0" and "pi" it is a single number. For "sigma" ind is ignored.}

\item{chains}{The indices of chains to be included in the plot. Default is -1,
which uses all the chains.}

\item{ylim}{The range of y-axis in the final plot}

\item{T}{The number of warmup period.}

\item{pars}{A parameter to be ploted. Must be one of "B", "sigma", "sigma1", 
"sigma0" and "pi".}
}
\value{
A traceplot of a scalar parameter after warmup-period
}
\description{
Produce traceplot for the specified parameter
}
