% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_functions.R
\name{zero.centered.em}
\alias{zero.centered.em}
\title{Perform EM Algorithm on Pleiotropy}
\usage{
zero.centered.em(
  values,
  std.errors,
  initial_slab_prob = 0.05,
  initial_sigma1_squared = 0.1,
  initial_sigma2_squared = 1e-04,
  iters = 100,
  optimizer = "Brent",
  show_outputs = FALSE
)
}
\arguments{
\item{values}{A vector of observations.}

\item{initial_slab_prob}{The initial value for the proportion of slab (the
Gaussian component with higher variance). Default is .05.}

\item{initial_sigma1_squared}{The initial value for the variance of the
slab component. Default is .1.}

\item{initial_sigma2_squared}{The initial value for the variance of the 
spike component. Default is .0001}

\item{iters}{The number of iterations of EM algorithm. Default is 100}

\item{optimizer}{The method used in Maximization Step. Must be one of "Brent",
"Nelder-Mead" and "L-BFGS-B". Default is "Brent".}

\item{show_outputs}{Whether to included a plot of the fitted density and the
smoothed data density. Default is FALSE.}

\item{std.error}{The standard error of noise of observations.}
}
\value{
A list of elements 
\item{S1} The standard deviation of the slab component
\item{S2} The standard deviation of the spike component
\item{Pi} The proportion of the slab component
}
\description{
Fit a two-component zero-mean Gaussian mixture on a sequence of observed 
values
}
